removed http adapter, the http package is multiplatform updated version number consolidated tests removed clientCompleter as it seems to be unused changed constructor parameters to be more clear moved test files up a level, no need for lib directory bumped version to 5.0.0 as this breaks backwards compatibility with 4.x
