+++ CHANGELOG.MD
@@ -1,3 +1,7 @@
+## 5.0.0
+Made multiplatform support cleaner by always using the multiplatform 'http' package
+You can now just import 'package:wilt/wilt.dart' into any project and it will Just Work(tm)
+
 ## 4.0.4
 Linter updates
