+++ analysis_options.yaml
@@ -5,7 +5,7 @@
 
 analyzer:
   exclude:
-  - test/issues/**
+    - test/issues/**
 
 linter:
   rules:
@@ -81,7 +81,6 @@ linter:
     - no_duplicate_case_values
     - non_constant_identifier_names
     - null_closures
-    - omit_local_variable_types
     - one_member_abstracts
     - only_throw_errors
     - overridden_fields
@@ -143,7 +142,6 @@ linter:
     - unnecessary_await_in_return
     - unnecessary_brace_in_string_interps
     - unnecessary_const
-    - unnecessary_final
     - unnecessary_getters_setters
     - unnecessary_lambdas
     - unnecessary_new
