+++ lib/wilt.dart
@@ -10,6 +10,7 @@ library wilt;
 import 'dart:async';
 import 'dart:convert';
 import 'package:json_object_lite/json_object_lite.dart' as jsonobject;
+import 'package:http/http.dart' as http;
 
 part 'src/wilt.dart';
 
@@ -22,5 +23,3 @@ part 'src/wilt_change_notification_parameters.dart';
 part 'src/wilt_change_notification_event.dart';
 
 part 'src/wilt_change_notification.dart';
-
-part 'src/httpAdapters/wilt_http_adapter.dart';
