+++ pubspec.yaml
@@ -1,18 +1,18 @@
 name: wilt
-version: 4.0.4
+version: 5.0.0
 description: A browser/server based CouchDB client library supporting attachments and change notification.
 homepage: https://github.com/shamblett/wilt
 
 environment:
-  sdk: '>=2.0.0-dev <3.0.0'
+  sdk: ">=2.0.0-dev <3.0.0"
 
 dependencies:
-  crypto: '^2.1.3'
-  http: '^0.12.0+2'
-  json_object_lite: '^2.1.0'
+  crypto: "^2.1.3"
+  http: "^0.12.1"
+  json_object_lite: "^2.1.0"
 
 dev_dependencies:
-  test: '^1.9.4'
-  build_runner: '^1.7.2'
-  build_test: '^0.10.10'
-  build_web_compilers: '^2.7.1'
+  test: "^1.9.4"
+  build_runner: "^1.7.2"
+  build_test: "^0.10.10"
+  build_web_compilers: "^2.7.1"
