--- test/lib/wilt_test_config.dart
@@ -1,16 +1,0 @@
-// CouchDB server
-const String hostName = 'localhost';
-const String port = '5984';
-const String serverPort = '5984';
-const String scheme = 'http://';
-
-// Databases to use for testing
-const String databaseNameBrowser = 'wilttest_browser';
-const String databaseNameServer = 'wilttest_server';
-
-/// Authentication, set as you wish, note leaving userName and password null
-/// implies no authentication, i.e admin party, if set Basic authentication is
-/// used.
-
-const String userName = 'steve';
-const String userPassword = 'setacrepes';
