+++ test/wilt_test_config.dart.default
@@ -1,0 +1,16 @@
+// Copy this file to wilt_test_config.dart and modify as needed for your testing environment
+
+// CouchDB server
+const String hostName = 'localhost';
+const int serverPort = 5984;
+const bool useSSL = false;
+
+// Databases to use for testing
+const String databaseName = 'wilt_test';
+
+/// Authentication, set as you wish, note leaving userName and password null
+/// implies no authentication, i.e admin party, if set Basic authentication is
+/// used.
+
+const String userName = 'admin';
+const String userPassword = 'admin';
